Fixing the document date for REC-1.2.
The original upload had 2022-11-02 as the document date, while it
was really uploaded on 2021-11-02, which is also the metadata in
the document repository.

To prevent potential confusion, this branch produces documents with the
right date.  To make things build, I also had to update a bibcode that
went from rept to spec in the meantime.